fix(vite): Update CJS to ESM for Vite
