+++ packages/ui/package.json
@@ -28,7 +28,7 @@
   ],
   "scripts": {
     "start": "vite",
-    "build": "tsc && vite build --config vite.config.cjs",
+    "build": "tsc && vite build --config vite.config.js",
     "build:lib": "yarn rimraf ./lib && yarn build:esm && yarn build:cjs",
     "build:esm": "tsc --project tsconfig.esm.json && copyfiles -u 1 ./src/**/*.scss ./src/assets/** ./lib/esm",
     "build:cjs": "tsc --project tsconfig.cjs.json && copyfiles -u 1 ./src/**/*.scss ./src/assets/** ./lib/cjs",
