+++ packages/ui/scripts/get-camel-catalog-files.js
@@ -1,11 +1,9 @@
-// eslint-disable-next-line no-undef, @typescript-eslint/no-var-requires
-const fs = require('node:fs');
-// eslint-disable-next-line no-undef, @typescript-eslint/no-var-requires
-const path = require('node:path');
-// eslint-disable-next-line no-undef, @typescript-eslint/no-var-requires
-const vite = require('vite');
-const { normalizePath } = vite;
+import { readdirSync } from 'node:fs';
+import { createRequire } from 'node:module';
+import { dirname, join } from 'node:path';
+import { normalizePath } from 'vite';
 
+const require = createRequire(import.meta.url);
 /**
  * Temporary function to copy the built Kaoto Camel Catalog into the assets folder
  *
@@ -13,22 +11,20 @@ const { normalizePath } = vite;
  * and this file can be restored to a .ts file.
  * Related issue: https://github.com/sveltejs/vite-plugin-svelte/issues/141#issuecomment-898900239
  */
-function getCamelCatalogFiles() {
+export const getCamelCatalogFiles = () => {
   let camelCatalogPath = '';
 
   try {
-    // eslint-disable-next-line no-undef
-    camelCatalogPath = normalizePath(path.dirname(require.resolve('@kaoto-next/camel-catalog/package.json')));
-    camelCatalogPath = path.join(camelCatalogPath, 'dist');
+    const camelCatalogIndexJsonPath = require.resolve('@kaoto-next/camel-catalog/index.json');
+    camelCatalogPath = normalizePath(dirname(camelCatalogIndexJsonPath));
   } catch (error) {
-    console.error(error);
-    /* empty */
+    throw new Error(`Could not find '@kaoto-next/camel-catalog' \n\n ${error}`);
   }
 
   console.info(`Found '@kaoto-next/camel-catalog' in ${camelCatalogPath}`, '\n');
 
   try {
-    if (fs.readdirSync(camelCatalogPath).length === 0) {
+    if (readdirSync(camelCatalogPath).length === 0) {
       throw new Error();
     }
   } catch (error) {
@@ -43,15 +39,9 @@ function getCamelCatalogFiles() {
   }
 
   /** List all the JSON files in the Camel Catalog folder */
-  const jsonFiles = fs
-    .readdirSync(camelCatalogPath)
+  const jsonFiles = readdirSync(camelCatalogPath)
     .filter((file) => file.endsWith('.json'))
-    .map((file) => path.join(camelCatalogPath, file));
+    .map((file) => join(camelCatalogPath, file));
 
   return jsonFiles;
-}
-
-// eslint-disable-next-line no-undef
-module.exports = {
-  getCamelCatalogFiles,
 };
