+++ packages/ui/vite.config.js
@@ -1,7 +1,7 @@
 import react from '@vitejs/plugin-react';
 import { defineConfig } from 'vite';
 import { viteStaticCopy } from 'vite-plugin-static-copy';
-import { getCamelCatalogFiles } from './scripts/get-camel-catalog-files.cjs';
+import { getCamelCatalogFiles } from './scripts/get-camel-catalog-files';
 
 // https://vitejs.dev/config/
 export default defineConfig({
